+++ lib/searchkick/query.rb
@@ -18,8 +18,8 @@ module Searchkick
       unknown_keywords = options.keys - [:aggs, :body, :body_options, :boost,
         :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :debug, :emoji, :exclude, :execute, :explain,
         :fields, :highlight, :includes, :index_name, :indices_boost, :limit, :load,
-        :match, :misspellings, :offset, :operator, :order, :padding, :page, :per_page, :profile,
-        :request_params, :routing, :select, :similar, :smart_aggs, :suggest, :slop, :track, :type, :where]
+        :match, :misspellings, :model_includes, :offset, :operator, :order, :padding, :page, :per_page, :profile,
+        :request_params, :routing, :select, :similar, :smart_aggs, :suggest, :track, :type, :where]
       raise ArgumentError, "unknown keywords: #{unknown_keywords.join(", ")}" if unknown_keywords.any?
 
       term = term.to_s
@@ -79,7 +79,7 @@ module Searchkick
       @execute ||= begin
         begin
           response = execute_search
-          if @misspellings_below && response["hits"]["total"] < @misspellings_below
+          if retry_misspellings?(response)
             prepare
             response = execute_search
           end
@@ -108,15 +108,13 @@ module Searchkick
         padding: @padding,
         load: @load,
         includes: options[:includes],
+        model_includes: options[:model_includes],
         json: !@json.nil?,
         match_suffix: @match_suffix,
         highlighted_fields: @highlighted_fields || [],
-        misspellings: @misspellings
+        misspellings: @misspellings,
+        term: term
       }
-
-      Rails.logger.warn "\n\n~~~> handle_response: #{response.inspect}"
-      Rails.logger.warn "\n\n~~~> opts: #{opts.inspect}"
-      Rails.logger.warn "\n\n~~~> debug?: #{options[:debug].inspect}"
 
       if options[:debug]
         require "pp"
@@ -163,6 +161,10 @@ module Searchkick
       @execute = Searchkick::Results.new(searchkick_klass, response, opts)
     end
 
+    def retry_misspellings?(response)
+      @misspellings_below && response["hits"]["total"] < @misspellings_below
+    end
+
     private
 
     def handle_error(e)
@@ -206,7 +208,6 @@ module Searchkick
       boost_fields, fields = set_fields
 
       operator = options[:operator] || "and"
-      slop = options[:slop] || 0
 
       # pagination
       page = [options[:page].to_i, 1].max
@@ -223,12 +224,17 @@ module Searchkick
 
       @json = options[:body]
       if @json
+        ignored_options = options.keys & [:aggs, :boost,
+          :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :exclude, :explain,
+          :fields, :highlight, :indices_boost, :limit, :match, :misspellings, :offset, :operator, :order,
+          :padding, :page, :per_page, :select, :smart_aggs, :suggest, :where]
+        warn "The body option replaces the entire body, so the following options are ignored: #{ignored_options.join(", ")}" if ignored_options.any?
         payload = @json
       else
         if options[:similar]
           payload = {
             more_like_this: {
-              like_text: term,
+              like: term,
               min_doc_freq: 1,
               min_term_freq: 1,
               analyzer: "searchkick_search2"
@@ -297,13 +303,12 @@ module Searchkick
               end
 
             shared_options[:operator] = operator if match_type == :match
-            shared_options[:slop] = slop if match_type == :match_phrase
 
             exclude_analyzer = nil
             exclude_field = field
 
             if field == "_all" || field.end_with?(".analyzed")
-              shared_options[:cutoff_frequency] = 0.001 unless operator == "and" || misspellings == false
+              shared_options[:cutoff_frequency] = 0.001 unless operator.to_s == "and" || misspellings == false
               qs.concat [
                 shared_options.merge(analyzer: "searchkick_search"),
                 shared_options.merge(analyzer: "searchkick_search2")
@@ -445,8 +450,14 @@ module Searchkick
         # indices_boost
         set_boost_by_indices(payload)
 
+        # type when inheritance
+        where = (options[:where] || {}).dup
+        if searchkick_options[:inheritance] && (options[:type] || (klass != searchkick_klass && searchkick_index))
+          where[:type] = [options[:type] || klass].flatten.map { |v| searchkick_index.klass_document_type(v, true) }
+        end
+
         # filters
-        filters = where_filters(options[:where])
+        filters = where_filters(where)
         set_filters(payload, filters) if filters.any?
 
         # aggregations
@@ -462,8 +473,7 @@ module Searchkick
         payload[:timeout] ||= "#{Searchkick.search_timeout + 1}s"
 
         # An empty array will cause only the _id and _type for each hit to be returned
-        # doc for :select - http://www.elasticsearch.org/guide/reference/api/search/fields/
-        # doc for :select_v2 - https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-source-filtering.html
+        # https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-source-filtering.html
         if options[:select]
           if options[:select] == []
             # intuitively [] makes sense to return no fields, but ES by default returns all fields
@@ -474,14 +484,15 @@ module Searchkick
         elsif load
           payload[:_source] = false
         end
-
-        if options[:type] || (klass != searchkick_klass && searchkick_index)
-          @type = [options[:type] || klass].flatten.map { |v| searchkick_index.klass_document_type(v) }
-        end
-
-        # routing
-        @routing = options[:routing] if options[:routing]
       end
+
+      # type
+      if !searchkick_options[:inheritance] && (options[:type] || (klass != searchkick_klass && searchkick_index))
+        @type = [options[:type] || klass].flatten.map { |v| searchkick_index.klass_document_type(v) }
+      end
+
+      # routing
+      @routing = options[:routing] if options[:routing]
 
       # merge more body options
       payload = payload.deep_merge(options[:body_options]) if options[:body_options]
@@ -495,7 +506,8 @@ module Searchkick
 
     def set_fields
       boost_fields = {}
-      fields = options[:fields] || searchkick_options[:searchable]
+      fields = options[:fields] || searchkick_options[:default_fields] || searchkick_options[:searchable]
+      all = searchkick_options.key?(:_all) ? searchkick_options[:_all] : below60?
       default_match = options[:match] || searchkick_options[:match] || :word
       fields =
         if fields
@@ -506,12 +518,14 @@ module Searchkick
             boost_fields[field] = boost.to_f if boost
             field
           end
-        elsif default_match == :word
+        elsif all && default_match == :word
           ["_all"]
-        elsif default_match == :phrase
+        elsif all && default_match == :phrase
           ["_all.phrase"]
+        elsif term == "*"
+          []
         else
-          raise ArgumentError, "Must specify fields"
+          raise ArgumentError, "Must specify fields to search"
         end
       [boost_fields, fields]
     end
@@ -836,7 +850,7 @@ module Searchkick
         if value.any?(&:nil?)
           {bool: {should: [term_filters(field, nil), term_filters(field, value.compact)]}}
         else
-          {in: {field => value}}
+          {terms: {field => value}}
         end
       elsif value.nil?
         {bool: {must_not: {exists: {field: field}}}}
@@ -877,13 +891,21 @@ module Searchkick
           }
         }
 
-        {
-          filter: {
+        if value[:missing]
+          if below50?
+            raise ArgumentError, "The missing option for boost_by is not supported in Elasticsearch < 5"
+          else
+            script_score[:field_value_factor][:missing] = value[:missing].to_f
+          end
+        else
+          script_score[:filter] = {
             exists: {
               field: field
             }
           }
-        }.merge(script_score)
+        end
+
+        script_score
       end
     end
 
@@ -910,6 +932,10 @@ module Searchkick
 
     def below50?
       Searchkick.server_below?("5.0.0-alpha1")
+    end
+
+    def below60?
+      Searchkick.server_below?("6.0.0-alpha1")
     end
   end
 end