+++ test/suggest_test.rb
@@ -69,12 +69,16 @@ class SuggestTest < Minitest::Test
 
   def test_multiple_models
     store_names ["Great White Shark", "Hammerhead Shark", "Tiger Shark"]
-    assert_equal "how big is a tiger shark", Searchkick.search("How Big is a Tigre Shar", suggest: [:name]).suggestions.first
+    assert_equal "how big is a tiger shark", Searchkick.search("How Big is a Tigre Shar", suggest: [:name], fields: [:name]).suggestions.first
   end
 
   def test_multiple_models_no_fields
     store_names ["Great White Shark", "Hammerhead Shark", "Tiger Shark"]
     assert_raises(ArgumentError) { Searchkick.search("How Big is a Tigre Shar", suggest: true) }
+  end
+
+  def test_star
+    assert_equal [], Product.search("*", suggest: true).suggestions
   end
 
   protected