+++ test/test_helper.rb
@@ -42,6 +42,10 @@ def elasticsearch_below50?
   Searchkick.server_below?("5.0.0-alpha1")
 end
 
+def elasticsearch_below60?
+  Searchkick.server_below?("6.0.0-alpha1")
+end
+
 def elasticsearch_below22?
   Searchkick.server_below?("2.2.0")
 end
@@ -94,14 +98,6 @@ if defined?(Mongoid)
     field :text
   end
 
-  class Topic
-    include Mongoid::Document
-
-    field :name
-    field :before_called, type: Boolean
-    field :after_called, type: Boolean
-  end
-
   class Speaker
     include Mongoid::Document
 
@@ -118,6 +114,12 @@ if defined?(Mongoid)
   end
 
   class Cat < Animal
+  end
+
+  class Sku
+    include Mongoid::Document
+
+    field :name
   end
 elsif defined?(NoBrainer)
   NoBrainer.configure do |config|
@@ -160,15 +162,6 @@ elsif defined?(NoBrainer)
     field :text, type: Text
   end
 
-  class Topic
-    include NoBrainer::Document
-
-    field :id,   type: Object
-    field :name, type: String
-    field :before_called, type: Boolean
-    field :after_called, type: Boolean
-  end
-
   class Speaker
     include NoBrainer::Document
 
@@ -187,6 +180,13 @@ elsif defined?(NoBrainer)
   end
 
   class Cat < Animal
+  end
+
+  class Sku
+    include NoBrainer::Document
+
+    field :id,   type: String
+    field :name, type: String
   end
 elsif defined?(Cequel)
   cequel =
@@ -267,6 +267,13 @@ elsif defined?(Cequel)
   end
 
   class Cat < Animal
+  end
+
+  class Sku
+    include Cequel::Record
+
+    key :id, :uuid
+    column :name, :text
   end
 
   [Product, Store, Region, Speaker, Animal].each(&:synchronize_schema)
@@ -356,10 +363,8 @@ else
     t.string :type
   end
 
-  ActiveRecord::Migration.create_table :topics do |t|
+  ActiveRecord::Migration.create_table :skus, id: :uuid do |t|
     t.string :name
-    t.boolean :after_called
-    t.boolean :before_called
   end
 
   class Product < ActiveRecord::Base
@@ -373,9 +378,6 @@ else
   class Region < ActiveRecord::Base
   end
 
-  class Topic < ActiveRecord::Base
-  end
-
   class Speaker < ActiveRecord::Base
   end
 
@@ -386,6 +388,9 @@ else
   end
 
   class Cat < Animal
+  end
+
+  class Sku < ActiveRecord::Base
   end
 end
 
@@ -443,6 +448,7 @@ end
 
 class Store
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
     routing: true,
     merge_mappings: true,
     mappings: {
@@ -464,6 +470,7 @@ end
 
 class Region
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
     geo_shape: {
       territory: {tree: "quadtree", precision: "10km"}
     }
@@ -479,32 +486,9 @@ class Region
   end
 end
 
-class Topic
-
-  searchkick \
-    before_index: :before_callback,
-    after_index: :after_callback
-
-  def search_data
-    {
-      name: name
-    }
-  end
-
-  protected
-
-  # This is extremely clunky, but still better than the stubbing support in minitest.
-  def before_callback
-    update_column :before_called, true
-  end
-
-  def after_callback
-    update_column :after_called, true
-  end
-end
-
 class Speaker
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
     conversions: ["conversions_a", "conversions_b"]
 
   attr_accessor :conversions_a, :conversions_b, :aisle
@@ -520,11 +504,17 @@ end
 
 class Animal
   searchkick \
+    default_fields: elasticsearch_below60? ? nil : [:name],
+    inheritance: !elasticsearch_below60?,
     text_start: [:name],
     suggest: [:name],
     index_name: -> { "#{name.tableize}-#{Date.today.year}#{Searchkick.index_suffix}" },
     callbacks: defined?(ActiveJob) ? :async : true
     # wordnet: true
+end
+
+class Sku
+  searchkick callbacks: defined?(ActiveJob) ? :async : true
 end
 
 Product.searchkick_index.delete if Product.searchkick_index.exists?
@@ -543,6 +533,7 @@ class Minitest::Test
     Store.destroy_all
     Animal.destroy_all
     Speaker.destroy_all
+    Sku.destroy_all
   end
 
   protected